move embeds.js to util
