+++ djs-bot/util/getChannel.js
@@ -1,6 +1,6 @@
 const Bot = require("../lib/Bot");
 const { MessageEmbed } = require("../lib/Embed");
-const { redEmbed } = require("../lib/embeds");
+const { redEmbed } = require("../util/embeds");
 
 // Module checks if you meet the channel requirements to use music commands
 /**
